requirements.txt
astor==0.8.1
attrdict==2.0.1
attrs==22.1.0
Babel==2.10.3
bce-python-sdk==0.8.74
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
cmake==3.24.1
colorama==0.4.5
concurrent-log-handler==0.9.19
cryptography==38.0.1
cssselect==1.1.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.30
DBUtils==3.0.2
decorator==4.4.2
distlib==0.3.6
elastic-transport==8.4.0
elasticsearch==8.4.1
et-xmlfile==1.1.0
filelock==3.8.0
flake8==5.0.4
Flask==2.2.2
Flask-Babel==2.0.0
fonttools==4.37.1
future==0.18.2
gast==0.3.3
identify==2.5.5
idna==3.3
imageio==2.21.2
imgaug==0.4.0
importlib-metadata==4.2.0
itsdangerous==2.1.2
jarowinkler==1.2.1
Jinja2==3.1.2
kafka-python==2.0.2
kiwisolver==1.4.4
lmdb==1.3.0
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.3
mccabe==0.7.0
nb-filelock==0.7
nb-log==7.7
networkx==2.6.3
nodeenv==1.7.0
numpy==1.19.3
opcua==0.98.13
opencv-contrib-python==4.4.0.46
opencv-python==4.5.5.64
openpyxl==3.0.10
packaging==21.3
paddleocr==2.5.0.3
paddlepaddle==2.1.0
pandas==1.3.5
pika==1.3.0
Pillow==9.2.0
platformdirs==2.5.2
portalocker==2.5.1
pre-commit==2.20.0
premailer==3.10.0
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post2
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.15.0
pyflakes==2.5.0
pymongo==4.0.2
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==0.1.10
python-Levenshtein==0.12.2
pytz==2022.2.1
PyWavelets==1.3.0
pywin32==304
PyYAML==6.0
pyzbar==0.1.9
rapidfuzz==2.6.1
requests==2.28.1
scikit-image==0.19.2
scipy==1.7.3
service-identity==21.1.0
Shapely==1.8.2
shellcheck-py==0.8.0.4
six==1.16.0
threadpool==1.3.2
tifffile==2021.11.2
toml==0.10.2
tomorrow3==1.1.0
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.16.2
visualdl==2.2.3
watchdog==2.1.9
Werkzeug==2.2.2
zipp==3.8.1
